change UserRecipe rating to integer

+++ app/graphql/types/user_recipe_type.rb
@@ -4,6 +4,6 @@ module Types
     field :user_id, Integer, null: false
     field :recipe_id, Integer, null: false
     field :amounted_donated, Float, null: false
-    field :recipe_rating, Float, null: false
+    field :recipe_rating, Integer, null: false
   end
 end
